Updated - 06/06

+++ 2024b.Tal.Mizrahi/src/main/java/demo/services/AdminServiceImplementation.java
@@ -3,6 +3,8 @@ package demo.services;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Sort.Direction;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import demo.boundaries.MiniAppCommandBoundary;
@@ -12,6 +14,8 @@ import demo.converters.UserConverter;
 import demo.crud.CommandCrud;
 import demo.crud.ObjectCrud;
 import demo.crud.UserCrud;
+import demo.entities.UserEntity;
+import demo.objects.RolesEnum;
 
 @Service
 public class AdminServiceImplementation implements AdminService {
@@ -43,7 +47,9 @@ public class AdminServiceImplementation implements AdminService {
 
 	@Override
 	@Transactional(readOnly = false)
-	public void deleteAllUsers() {
+	public void deleteAllUsers(String userSuperapp, String email) {
+		
+		checkAdminPermission(userSuperapp, email);
 		this.userCrud.deleteAll();
 		System.err.println("All user entries Deleted");
 		
@@ -51,7 +57,8 @@ public class AdminServiceImplementation implements AdminService {
 
 	@Override
 	@Transactional(readOnly = false)
-	public void deleteAllObjects() {
+	public void deleteAllObjects(String userSuperapp, String email) {
+		checkAdminPermission(userSuperapp, email);
 		this.objectCrud.deleteAll();
 		System.err.println("All object entries Deleted");
 		
@@ -59,7 +66,8 @@ public class AdminServiceImplementation implements AdminService {
 
 	@Override
 	@Transactional(readOnly = false)
-	public void deleteAllCommandsHistory() {
+	public void deleteAllCommandsHistory(String userSuperapp, String email) {
+		checkAdminPermission(userSuperapp, email);
 		this.commandCrud.deleteAll();
 		System.err.println("All commands entries Deleted");
 		
@@ -67,8 +75,9 @@ public class AdminServiceImplementation implements AdminService {
 
 	@Override
 	@Transactional(readOnly = true)
-	public List<UserBoundary> getAllUsers() {
-		return this.userCrud.findAll() // List<UserEntity>
+	public List<UserBoundary> getAllUsers(String userSuperapp, String email, int size, int page) {
+		checkAdminPermission(userSuperapp, email);
+		return this.userCrud.findAll(PageRequest.of(page, size, Direction.ASC, "role", "userId")) // List<UserEntity>
 				.stream() // Stream<DemoEntity>
 				.peek(entity -> System.err.println("* " + entity)) // Prints all items.
 				.map(userConverter::toBoundary) // Stream<userBoundary>
@@ -77,8 +86,9 @@ public class AdminServiceImplementation implements AdminService {
 
 	@Override
 	@Transactional(readOnly = true)
-	public List<MiniAppCommandBoundary> getAllCommands() {
-		return this.commandCrud.findAll() // List<UserEntity>
+	public List<MiniAppCommandBoundary> getAllCommands(String userSuperapp, String email, int size, int page) {
+		checkAdminPermission(userSuperapp, email);
+		return this.commandCrud.findAll(PageRequest.of(page, size, Direction.DESC, "commandId", "invocationTimesTamp")) // List<UserEntity>
 				.stream() // Stream<DemoEntity>
 				.peek(entity -> System.err.println("* " + entity)) // Prints all items.
 				.map(this.commandConverter::toBoundary) // Stream<CommandBoundary>
@@ -87,13 +97,25 @@ public class AdminServiceImplementation implements AdminService {
 
 	@Override
 	@Transactional(readOnly = true)
-	public List<MiniAppCommandBoundary> getCommandsOfSpecificMiniApp(String miniAppName) {
+	public List<MiniAppCommandBoundary> getCommandsOfSpecificMiniApp(String miniAppName, String userSuperapp, String email, int page, int size) {
+		checkAdminPermission(userSuperapp, email);
 		return commandCrud
-				.findAllByMiniAppName(miniAppName)
+				.findAllByMiniAppName(miniAppName, PageRequest.of(page, size, Direction.ASC, "command"))
 				.stream()
 				.map(commandConverter::toBoundary)
 				.peek(System.err :: println)
 				.toList();
+	}
+	
+	public void checkAdminPermission(String userSuperapp, String email) {
+		String id = userSuperapp + "#" + email;
+		UserEntity entity = this.userCrud
+				.findById(id)
+				.orElseThrow(() -> new NotFoundException("UserEntity with email: " + email 
+						+ " and superapp " + userSuperapp + " Does not exist in database"));
+		if (!entity.getRole().equals(RolesEnum.ADMIN))
+			throw new ForbiddenException("UserEntity with email: " + email 
+						+ " and superapp " + userSuperapp + " not have the permission");
 	}
 
 }
